backend/requirements.txt
absl-py==0.12.0
astor==0.8.1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
Django==2.2.20
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.4
gast==0.2.2
google-auth==1.23.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.37.1
h5py==2.10.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
importlib-metadata==2.1.1
joblib==0.14.1
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.2.2
matplotlib==3.0.3
moviepy==1.0.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==0.25.3
Pillow==7.2.0
proglog==0.1.9
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.25.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.9.1
six==1.15.0
sklearn==0.0
SpeechRecognition==3.8.1
sqlparse==0.4.1
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
tqdm==4.60.0
urllib3==1.26.4
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==1.2.0
